Revert "Modularize routes"
