--- api/src/routes/User/login.js
@@ -1,37 +1,0 @@
-"use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-const express_1 = require("express");
-const User_1 = __importDefault(require("../../models/User"));
-const bcrypt_1 = __importDefault(require("bcrypt"));
-const router = (0, express_1.Router)();
-router.post("/", (req, res) => __awaiter(void 0, void 0, void 0, function* () {
-    try {
-        const { email, password } = req.body;
-        const user = yield User_1.default.findOne({ email });
-        if (!user)
-            return res.status(400).send('Usuario inexistente');
-        const passwordCompare = yield bcrypt_1.default.compare(password, user.password);
-        if (passwordCompare) {
-            const { email, userName, lastName, avatar, Account } = user;
-            const token = user.generateAuthToken();
-            return res.cookie("access_token", token, { maxAge: 7 * 24 * 3600 * 1000, httpOnly: true }).status(200).send({ email, userName, lastName, avatar, Account });
-        }
-        res.status(400).send('Contraseña Incorrecta');
-    }
-    catch (err) {
-        res.status(404).send(err);
-    }
-}));
-exports.default = router;
