+++ api/src/routes/index.js
@@ -1,13 +1,13 @@
 "use strict";
-// Backup:
-// import routeUserNoSql from '../../backup/routeUserNoSql'
-// router.use('/test', routeUserNoSql)
 var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const express_1 = require("express");
-const User_1 = __importDefault(require("./User"));
 const router = (0, express_1.Router)();
-router.use("/", User_1.default);
+// Backup:
+// import routeUserNoSql from '../../backup/routeUserNoSql'
+// router.use('/test', routeUserNoSql)
+const routesUser_1 = __importDefault(require("./routesUser"));
+router.use("/", routesUser_1.default);
 exports.default = router;
