+++ api/src/routes/index.ts
@@ -1,13 +1,12 @@
+import { Router } from "express";
+
+ const router = Router()
 // Backup:
 // import routeUserNoSql from '../../backup/routeUserNoSql'
 // router.use('/test', routeUserNoSql)
 
+import routesUser from './routesUser';
 
-import { Router } from "express";
-import User from './User';
-
-const router = Router()
-
-router.use("/", User)
+ router.use("/", routesUser)
 
 export default router
