+++ client/tsconfig.json
@@ -20,6 +20,7 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react-jsx",
+    "types": ["google"]
   },
   "include": [
     "src"
